[modify] 랭킹 구현중

+++ runnershi_be/src/main/java/com/ssafy/runnershi/repository/AlarmRepository.java
@@ -3,6 +3,6 @@ package com.ssafy.runnershi.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.ssafy.runnershi.entity.Alarm;
 
-public interface AlarmRepository extends JpaRepository<Alarm, Integer> {
+public interface AlarmRepository extends JpaRepository<Alarm, Long> {
 
 }
